Move some helper from nnapi ep to shared
