+++ onnxruntime/core/providers/nnapi/nnapi_builtin/builders/op_builder.cc
@@ -7,6 +7,7 @@
 #include <core/providers/common.h>
 #include <onnx/onnx_pb.h>
 
+#include "core/providers/shared/utils.h"
 #include "helper.h"
 #include "model_builder.h"
 #include "op_builder.h"
